+++ packages/react-core/src/utils/geo.d.ts
@@ -1,0 +1,6 @@
+import { Viewport } from '../types';
+import { Polygon, MultiPolygon } from 'geojson';
+
+export function getGeometryToIntersect(viewport: Viewport | null, geometry: Polygon | MultiPolygon | null): Polygon | MultiPolygon | null;
+
+export function isGlobalViewport(viewport: Viewport | null): boolean;