+++ packages/react-redux/src/slices/cartoSlice.js
@@ -337,6 +337,13 @@ export const checkIfSourceIsDroppingFeature = (state, id) =>
   state.carto.dataSources[id]?.isDroppingFeatures;
 
 /**
+ * Redux selector to select the active viewport
+ */
+export const selectViewport = (state) => {
+  return state.carto.viewport ? state.carto.viewport : null;
+};
+
+/**
  * Redux selector to select the spatial filter of a given sourceId or the root one
  */
 export const selectSpatialFilter = (state, sourceId) => {
@@ -373,7 +380,9 @@ const NOT_ALLOWED_DECK_PROPS = [
 ];
 
 /**
- * Action to set the current ViewState
+ * Action to set the current ViewState.
+ *
+ * Requires redux-thunk middleware, also invokes debounced `setViewPort`.
  * @param {Object} viewState
  */
 export const setViewState = (viewState) => {
@@ -393,6 +402,26 @@ export const setViewState = (viewState) => {
     debouncedSetViewPort(dispatch, _setViewPort);
   };
 };
+
+/**
+ * Action to set the current ViewState.
+ *
+ * Doesn't refresh widgets immetiately, requires user to call `setViewPort` once all updates are ready.
+ * @param {Object} viewState
+ */
+export const setViewStateDirect = (viewState) => ({
+  type: 'carto/setViewState',
+  payload: viewState
+});
+
+/**
+ * Sync current viewport state deriving it from `viewState`.
+ *
+ * Causes widgets in remote mode to refresh its data.
+ */
+export const setViewPort = () => ({
+  type: 'carto/setViewPort'
+});
 
 /**
  * Action to set the ready features state of a layer